Create TheseusMovement
This is a temp class that in the future,  when the rest of the classes come along can stand alone ( with extra additions ofc ) or merged with Theseas.java. Lots to be decided such as speed , the actual labyrinth ( and IMPORTANT its numerical borders ), screen size etc. TBH algorithm not completed understood. But, basically, this going to be Theseus' movement.